pyproject.toml
[tool.poetry]
name = "bookmarks-service"
version = "0.1.0"
description = "REST API backend with a service for storing bookmarks"
authors = ["Ilya Malashenko <melax08@yandex.ru>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "bookmarks_service"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
djangorestframework = "3.14.0"
djoser = "2.2.2"
pillow = "10.1.0"
python-dotenv = "1.0.0"
redis = "5.0.1"
metadata-parser = "0.12.0"
psycopg2-binary = "2.9.6"
gunicorn = "20.1.0"
drf-yasg = "1.21.7"
celery = "5.3.6"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
